Fixed some bugs in Malika's MVC model
Added variables at the top that define the width and height of the screen, and replaced all declarations of width and height throughout the view.py with the variable names so that only top location needs to be updated if dimensions must be changed in the future.

Added variables (char_width, char_height) that get the width and height of the input image (spaceship) and plugs that into ship declaration (when making a new player object). This accounts for if the ship image is changed in the future.

Also added left and right bar for displaying
leader board, health, wallet, and store.

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Still to be implemented in MVC : demoFunctions.py