Made Create Text Function.
Then Created LeaderBoard and Category Titles and blitted them to specific locations.

Commmented out : Is a for loop that will go over a hypothetical sorted List.

What we want to use the loop for is to take each object from loop create text of each user, kills, deaths and create text for placing and blit it under each location.

commented Out is FOR LOOP for leader board